+++ src/seventh/client/gfx/InGameOptionsDialogView.java
@@ -53,7 +53,7 @@ public class InGameOptionsDialogView implements Renderable {
 	 * @see seventh.client.gfx.Renderable#render(seventh.client.gfx.Canvas, seventh.client.gfx.Camera, long)
 	 */
 	@Override
-	public void render(Canvas canvas, Camera camera, long alpha) {
+	public void render(Canvas canvas, Camera camera, float alpha) {
 		
 		if(dialog.getWeaponDialog().isVisible()) {
 			weaponClassDialogView.render(canvas, camera, alpha);
