+++ src/seventh/ui/view/TextBoxView.java
@@ -68,7 +68,7 @@ public class TextBoxView implements Renderable {
 	/* (non-Javadoc)
 	 * @see org.myriad.render.Renderable#render(org.myriad.render.Renderer, org.myriad.render.Camera, org.myriad.core.TimeUnit)
 	 */
-	public void render(Canvas renderer, Camera camera, long alpha) {
+	public void render(Canvas renderer, Camera camera, float alpha) {
 		if ( this.textBox.isVisible() ) {			
 			
 			Rectangle bounds = textBox.getBounds();
@@ -116,7 +116,7 @@ public class TextBoxView implements Renderable {
 		}
 	}
 
-	private void renderGradiantBackground(Widget w, Canvas renderer, Camera camera, long alpha) {
+	private void renderGradiantBackground(Widget w, Canvas renderer, Camera camera, float alpha) {
 //		int gradiant = w.getGradiantColor();
 //		int bg = w.getBackgroundColor();
